mp misc: Address build failures and warnings
Changes required to build with RTEMS_MULTIPROCESSING and
associated tests without warnings.

+++ contrib/cpukit/utf8proc/utf8proc.c
@@ -392,7 +392,7 @@ ssize_t utf8proc_decompose(
       if (decomp_result < 0) return decomp_result;
       wpos += decomp_result;
       /* prohibiting integer overflows due to too long strings: */
-      if (wpos < 0 || wpos > SSIZE_MAX/sizeof(int32_t)/2)
+      if (wpos < 0 || wpos > (ssize_t) (SSIZE_MAX/sizeof(int32_t)/2))
         return UTF8PROC_ERROR_OVERFLOW;
     }
   }
